Producer/requirements.txt
kafka-python==2.0.2
requests==2.22.0